initial snake

+++ lab3-sw/hello.c
@@ -425,10 +425,10 @@ int main()
         for(int r = 0; r < 30; r++, offset+=40){
             for(int c = 0; c < 40; c+=4){
                 //get(map, (Key){0, 0})
-                int sprite1 = get(screen_map, (Key){c, r});
-                int sprite2 = get(screen_map, (Key){c+1, r});
-                int sprite3 = get(screen_map, (Key){c+2, r});
-                int sprite4 = get(screen_map, (Key){c+3, r});
+                unsigned short int sprite1 = get(screen_map, (Key){c, r});
+                unsigned short int sprite2 = get(screen_map, (Key){c+1, r});
+                unsigned short int sprite3 = get(screen_map, (Key){c+2, r});
+                unsigned short int sprite4 = get(screen_map, (Key){c+3, r});
                 if(sprite1 > 0 ){
                     printf("sprite1 %d\n", sprite1);
                 }
